Rework JedisSentinelPool to recover if it can't find a master initially
JedisSentinelPool used to expect to find a master immediately on
startup. This both meant it'd block on instantiation and that if it
couldn't connect to a sentinel, or if a sentinel didn't know a master,
it would throw an exception and never recover. Now it doesn't return
jedis connections while it can't find a master, but it will recover
if the sentinels ever recover.

This also fixes a bug where if a sentinel connection went down while
a switch-master occurred, a new master would be missed.